+++ src/jp_helpers.js
@@ -307,14 +307,27 @@ const JUPYTER_URL_PATH = './_jupyter_url.txt';
 
 const RUN_JUPYTER_SERVER_PYTHON_SCRIPT = `
 
+'''
+Run a Jupyter server and save the token URL in a file.
+This script is left here for testing and maintenance.
+The script is available as a string constant in '../src/jp_helpers.js'.
+'''
+
 # https://stackoverflow.com/questions/2804543/read-subprocess-stdout-line-by-line
 
 import subprocess
 import os
+import signal
+
+proc = None
+cmd = None
 
 JUPYTER_URL_PATH = '${JUPYTER_URL_PATH}'
 
 def run():
+    global proc, cmd
+    signal.signal(signal.SIGINT, exit_cleanly)
+    signal.signal(signal.SIGTERM, exit_cleanly)
     cmd = ['jupyter', 'notebook', '--port=3000', '--no-browser']
     print ('Starting jupyter server: ' + repr(cmd))
     proc = subprocess.Popen(cmd, stderr=subprocess.PIPE)
@@ -334,8 +347,13 @@ def run():
                     print ('Jupyter url emitted: ' + repr(sline))
                     print ('   saved to: ' + repr(JUPYTER_URL_PATH))
     finally:
-        print ('Stopping jupyter server', cmd)
-        proc.kill()
+        exit_cleanly()
+
+def exit_cleanly(*arguments):
+    print ('Stopping jupyter server', cmd)
+    proc.kill()
+    if os.path.exists(JUPYTER_URL_PATH):
+        print('removing', JUPYTER_URL_PATH)
         os.remove(JUPYTER_URL_PATH)
 
 if __name__=='__main__':
